Create README.md
Tribute page
Yes, tribute pages are a real thing. If you Google “tribute page,” you will find a comprehensive list of links showing you how to build tribute pages. Essentially a tribute page is a webpage dedicated in honor of someone you love, admire, or respect. It can be a person or a beloved pet. 

A tribute page is a perfect project for sharpening your HTML and CSS skills and knowledge. In this project, you will make a webpage where you can write and dedicate a tribute to someone and publish the same. Apart from the writeup for the tribute, you need to add relevant images, links, etc., on the page.